fix(cosmos3): resolve Hub-ID checkpoints for meta-init; keep timesteps continuous
Two post-push review findings (Codex):

- The default recipes (nvidia/Cosmos3-Nano + meta_init_transformer) joined the
  raw Hub ID into the sharded loader's weights dir, which only accepts local
  directories — the default config crashed at weight-load time. from_config now
  resolves a non-local path via snapshot_download(allow_patterns=
  ['transformer/*']) first (the wan21 VAE precedent); local paths and the eager
  path are untouched.
- Revert the int64 timestep truncation to the original continuous fp32
  sigma * num_train_timesteps: NVIDIA's official training (cosmos-framework
  omni_mot_model.py) keeps continuous fp32 timesteps, and training sigmas are
  logit-normal draws that never land on the inference grid anyway, so flooring
  only quantized the conditioning and biased it by -0.5 units on average. The
  diffusers-0.39 inference-side int64 truncation is documented in the README as
  a tolerated quirk, not a contract to align training to.

Verified: mocked from_config drive (hub-id+meta resolves via snapshot_download
with transformer/* patterns; local-dir+meta skips it; hub-id+eager untouched);
pipeline timestep block byte-matches the pre-truncation original; pre-commit
--all-files green.

+++ unirl/models/cosmos3/pipeline.py
@@ -158,21 +158,19 @@ class Cosmos3JointStage:
             compute_dtype=self.bundle.dtype,
             **action_kwargs,
         )
-        # Inference conditions on int64-truncated scheduler timesteps (set_timesteps
-        # casts every branch via .astype(np.int64); the pipeline fills torch.full with
-        # t.item()); mirror that exactly instead of the fractional sigma*T.
-        timestep = int(float(sigma.item()) * float(self.pipe.scheduler.config.num_train_timesteps))
+        # Continuous fp32 sigma*T like official Cosmos training; README: timestep conditioning.
+        timestep = float(sigma.item()) * float(self.pipe.scheduler.config.num_train_timesteps)
         kwargs["vision_timesteps"] = torch.full(
             (meta["num_noisy_vision_tokens"],),
             timestep,
-            dtype=torch.int64,
+            dtype=torch.float32,
             device=x0.device,
         )
         if actions is not None:
             kwargs["action_timesteps"] = torch.full(
                 (meta["num_noisy_action_tokens"],),
                 timestep,
-                dtype=torch.int64,
+                dtype=torch.float32,
                 device=x0.device,
             )
